requirements.txt
click==8.1.3
document==1.0
Flask==2.2.2
Flask-WTF==1.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
Pillow==9.3.0
python-docx==0.8.11
python-pptx==0.6.21
Werkzeug==2.2.2
WTForms==3.0.1
XlsxWriter==3.0.3
